// src/global/types/Places.ts
export enum PlaceType {
  MUSEUM = 'MUSEUM',
  LIBRARY = 'LIBRARY',
  ARCHIVE = 'ARCHIVE',
  RESTAURANT = 'RESTAURANT',
  ATTRACTION = 'ATTRACTION',
  PUBLIC_PLACE = 'PUBLIC_PLACE',
  OTHER = 'OTHER',
}

/**
 * @typedef {Object} PlaceTag
 * @property {string} NOLOSAY - The nolosay tag.
 * @property {string} DISABILITY_FRIENDLY - The disability friendly tag.
 * @property {string} DEAF_FRIENDLY - The deaf friendly tag.
 * @property {string} BLIND_FRIENDLY - The blind friendly tag.
 * @property {string} OTHER - The other tag.
 */
// eslint-disable-next-line no-shadow
export enum PlaceTag {
  NOLOSAY = 'NOLOSAY',
  DISABILITY_FRIENDLY = 'DISABILITY_FRIENDLY',
  DEAF_FRIENDLY = 'DEAF_FRIENDLY',
  BLIND_FRIENDLY = 'BLIND_FRIENDLY',
  OTHER = 'OTHER',
}
// ...
export interface Place {
  id: number
  name: string
  shortDescription: string
  longDescription: string
  pictures: {
    id: number
    hostingUrl: string
  }[]
  address: {
    id: number
    houseNumber: string
    street: string
    zip: string
    city: {
      id: number
      name: string
      zip: string
      department: {
        id: number
        name: string
        country: {
          id: number
          name: string
        }
      }
    }
    otherDetails: string
    latitude: number
    longitude: number
  }
  telNumber: string
  email: string
  website: string
  type: PlaceType
  price: number
  exhibitions: any[]
  tags: PlaceTag[]
}